requirements.txt
absl-py==2.4.0
blinker==1.9.0
branca==0.8.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
folium==0.20.0
geographiclib==2.1
geopy==2.4.1
idna==3.11
immutabledict==4.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.3
openpyxl==3.1.5
ortools==9.15.6755
pandas==3.0.1
protobuf==6.33.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
xyzservices==2025.11.0
